requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

sanic==0.5.4
Pillow==4.1.1
python-magic==0.4.13

click==6.7
requests==2.14.2

aiohttp==2.1.0
async_timeout==1.2.1
